+++ CHANGELOG.md
@@ -1,3 +1,24 @@
+## [0.10.6](https://github.com/molgenis/molgenis-app-lifelines-webshop/compare/v0.10.5...v0.10.6) (2019-11-13)
+
+
+### Bug Fixes
+
+* fix [#139](https://github.com/molgenis/molgenis-app-lifelines-webshop/issues/139) remove call to removed mutation ([7eb8e29](https://github.com/molgenis/molgenis-app-lifelines-webshop/commit/7eb8e29))
+
+## [0.10.5](https://github.com/molgenis/molgenis-app-lifelines-webshop/compare/v0.10.4...v0.10.5) (2019-11-13)
+
+
+### Bug Fixes
+
+* [#136](https://github.com/molgenis/molgenis-app-lifelines-webshop/issues/136) apply lifelines config defaults ([#137](https://github.com/molgenis/molgenis-app-lifelines-webshop/issues/137)) ([5ecf031](https://github.com/molgenis/molgenis-app-lifelines-webshop/commit/5ecf031))
+
+## [0.10.4](https://github.com/molgenis/molgenis-app-lifelines-webshop/compare/v0.10.3...v0.10.4) (2019-11-11)
+
+
+### Bug Fixes
+
+* custom formatter to deal with -1 count ([6e32441](https://github.com/molgenis/molgenis-app-lifelines-webshop/commit/6e32441)), closes [#131](https://github.com/molgenis/molgenis-app-lifelines-webshop/issues/131)
+
 ## [0.10.3](https://github.com/molgenis/molgenis-app-lifelines-webshop/compare/v0.10.2...v0.10.3) (2019-11-07)